Crate miden_air

Modules

ace

ACE circuit integration for ProcessorAir.

config

STARK configuration factories for different hash functions.

debug

Debug constraint checker for lifted AIRs. Debug constraint checker for lifted AIRs.

trace

Structs

AirWitness

Prover witness: trace matrix, public values, and variable-length public inputs.

Felt

A Felt backed by Plonky3’s Goldilocks field element.

ProcessorAir

Miden VM Processor AIR implementation.

PublicInputs

Enums

DeserializationError

Defines errors which can occur during deserialization.

Constants

NUM_PUBLIC_VALUES

Number of fixed-length public values for the Miden VM AIR.

Traits

AirBuilder

A builder which contains both a trace on which AIR constraints can be evaluated as well as a method of accumulating the AIR constraint evaluations.

AuxBuilder

Builder for constructing the auxiliary trace from a main trace and challenges.

BaseAir

The underlying structure of an AIR.

ByteReader

Defines how primitive values are to be read from Self.

ByteWriter

Defines how primitive values are to be written into Self.

Deserializable

Defines how to deserialize Self from bytes.

ExtensionBuilder

Extension of AirBuilder for working over extension fields.

LiftedAir

Super-trait for AIR definitions used by the lifted STARK prover/verifier.

LiftedAirBuilder

Super-trait bundling all builder capabilities needed by the lifted STARK system.

PermutationAirBuilder

Trait for builders supporting permutation arguments (e.g., for lookup constraints).

Serializable

Defines how to serialize Self into bytes.

ToElements