miden_air

Module trace

Modules

bus_message

Bus message coefficient indices.

chiplets

decoder

log_precompile

range

stack

Structs

Challenges

Encodes multiset/LogUp contributions as alpha + <beta, message>.

MainTrace

MainTraceRow

Column layout of the main trace row.

RowIndex

A newtype wrapper around a usize value representing a step in the execution trace.

Enums

RowIndexError

Represents the types of errors that can occur when converting from and into RowIndex and using its operations.

Constants

ACE_CHIPLET_WIRING_BUS_OFFSET

ACE chiplet wiring bus.

ACE_CHIPLET_WIRING_BUS_RANGE

ACE_CHIPLET_WIRING_BUS_WIDTH

AUX_TRACE_RAND_CHALLENGES

Number of random challenges used for auxiliary trace constraints.

AUX_TRACE_WIDTH

Auxiliary trace segment width.

CHIPLETS_BUS_AUX_TRACE_OFFSET

Chiplets bus auxiliary columns.

CHIPLETS_BUS_AUX_TRACE_RANGE

CHIPLETS_BUS_AUX_TRACE_WIDTH

CHIPLETS_OFFSET

CHIPLETS_RANGE

CHIPLETS_WIDTH

CHIPLET_S0_COL_IDX

CHIPLET_S1_COL_IDX

CHIPLET_S2_COL_IDX

CHIPLET_S3_COL_IDX

CHIPLET_S4_COL_IDX

CHIPLET_SELECTORS_RANGE

Shared chiplet selector columns at the start of the chiplets segment.

CLK_COL_IDX

CTX_COL_IDX

DECODER_AUX_TRACE_OFFSET

Decoder auxiliary columns

DECODER_AUX_TRACE_RANGE

DECODER_AUX_TRACE_WIDTH

DECODER_TRACE_OFFSET

DECODER_TRACE_RANGE

DECODER_TRACE_WIDTH

FN_HASH_OFFSET

FN_HASH_RANGE

HASHER_AUX_TRACE_RANGE

HASHER_AUX_TRACE_WIDTH

HASH_KERNEL_VTABLE_AUX_TRACE_OFFSET

Chiplets virtual table auxiliary column.

LOG_PRECOMPILE_LABEL

Label for log_precompile transcript state messages on the virtual table bus.

MAX_MESSAGE_WIDTH

Maximum number of coefficients used in bus message encodings.

MIN_TRACE_LEN

The minimum length of the execution trace. This is the minimum required to support range checks.

PADDED_TRACE_WIDTH

RANGE_CHECK_AUX_TRACE_OFFSET

Range check auxiliary columns

RANGE_CHECK_AUX_TRACE_RANGE

RANGE_CHECK_AUX_TRACE_WIDTH

RANGE_CHECK_TRACE_OFFSET

RANGE_CHECK_TRACE_RANGE

RANGE_CHECK_TRACE_WIDTH

STACK_AUX_TRACE_OFFSET

Stack auxiliary columns

STACK_AUX_TRACE_RANGE

STACK_AUX_TRACE_WIDTH

STACK_TRACE_OFFSET

STACK_TRACE_RANGE

STACK_TRACE_WIDTH

SYS_TRACE_OFFSET

SYS_TRACE_RANGE

SYS_TRACE_WIDTH

TRACE_WIDTH

Traits

AuxBuilder

Builder for constructing the auxiliary trace from a main trace and challenges.