miden_air::trace

Module chiplets

Modules

ace

bitwise

hasher

Hasher chiplet trace constants and types.

kernel_rom

memory

Constants

BITWISE_A_COL_IDX

The index within the main trace of the bitwise column holding the aggregated value of input a.

BITWISE_A_COL_RANGE

The index range within the main trace for the bit decomposition of a for bitwise operations.

BITWISE_B_COL_IDX

The index within the main trace of the bitwise column holding the aggregated value of input b.

BITWISE_B_COL_RANGE

The index range within the main trace for the bit decomposition of b for bitwise operations.

BITWISE_OUTPUT_COL_IDX

The index within the main trace of the bitwise column containing the aggregated output value.

BITWISE_PREV_OUTPUT_COL_IDX

The index within the main trace of the bitwise column containing the aggregated output value of the previous row.

BITWISE_SELECTOR_COL_IDX

The index within the main trace of the bitwise column containing selector indicating the type of bitwise operation (AND or XOR)

BITWISE_TRACE_OFFSET

The first column of the bitwise chiplet.

BITWISE_TRACE_RANGE

The column index range for the main trace of the bitwise column

HASHER_CAPACITY_COL_RANGE

The range of columns in the execution trace that contains the capacity portion of the hasher state.

HASHER_NODE_INDEX_COL_IDX

The index of the hasher’s node index column in the execution trace.

HASHER_RATE_COL_RANGE

The range of columns in the execution trace that contains the rate portion of the hasher state.

HASHER_SELECTOR_COL_RANGE

The column index range in the execution trace containing the selector columns in the hasher.

HASHER_STATE_COL_RANGE

The range of columns in the execution trace that contain the hasher’s state.

HASHER_TRACE_OFFSET

The first column of the hash chiplet.

MEMORY_CLK_COL_IDX

The index within the main trace of the column containing the clock cycle of the memory access.

MEMORY_CTX_COL_IDX

The index within the main trace of the column containing the memory context.

MEMORY_D0_COL_IDX

The column index within the main trace for the lower 16-bits of the delta between two consecutive memory context IDs, addresses, or clock cycles.

MEMORY_D1_COL_IDX

The column index within the main trace for the upper 16-bits of the delta between two consecutive memory context IDs, addresses, or clock cycles.

MEMORY_D_INV_COL_IDX

The column index within the main trace for the inverse of the delta between two consecutive memory context IDs, addresses, or clock cycles, used to enforce that changes are correctly constrained.

MEMORY_FLAG_SAME_CONTEXT_AND_WORD

Column to hold the flag indicating whether the current memory operation is in the same context and same word as the previous operation.

MEMORY_IDX0_COL_IDX

The index within the main trace of the column containing the 0’th memory index.

MEMORY_IDX1_COL_IDX

The index within the main trace of the column containing the 1st memory index.

MEMORY_IS_READ_COL_IDX

The index within the main trace of the column containing the memory read/write column.

MEMORY_IS_WORD_ACCESS_COL_IDX

The index within the main trace of the column containing the memory element/word column.

MEMORY_TRACE_OFFSET

The first column of the memory chiplet.

MEMORY_V_COL_RANGE

The column index range within the main trace which holds the memory value elements.

MEMORY_WORD_COL_IDX

The index within the main trace of the column containing the memory address.

NUM_ACE_SELECTORS

The number of columns in the chiplets which are used as selectors for the ACE chiplet.

NUM_BITWISE_SELECTORS

The number of columns in the chiplets which are used as selectors for the bitwise chiplet.

NUM_HASHER_SELECTORS

The number of columns in the chiplets which are used as selectors for the hasher chiplet.

NUM_KERNEL_ROM_SELECTORS

The number of columns in the chiplets which are used as selectors for the kernel ROM chiplet.

NUM_MEMORY_SELECTORS

The number of columns in the chiplets which are used as selectors for the memory chiplet.