miden_air::trace::chiplets

Module bitwise

Constants

A_COL_IDX

The index of the column holding the aggregated value of input a within the bitwise chiplet execution trace.

A_COL_RANGE

The index range for the bit decomposition of a within the bitwise chiplet’s trace.

BITWISE_AND

Specifies a bitwise AND operation.

BITWISE_AND_LABEL

Unique label computed as 1 plus the full chiplet selector with the bits reversed. selector = [1, 0 | 0], flag = rev(selector) + 1 = [0 | 0, 1] + 1 = 2

BITWISE_XOR

Specifies a bitwise XOR operation.

BITWISE_XOR_LABEL

Unique label computed as 1 plus the full chiplet selector with the bits reversed. selector = [1, 0 | 1], flag = rev(selector) + 1 = [1 | 0, 1] + 1 = 6

B_COL_IDX

The index of the column holding the aggregated value of input b within the bitwise chiplet execution trace.

B_COL_RANGE

The index range for the bit decomposition of b within the bitwise chiplet’s trace.

NUM_DECOMP_BITS

The number of bits decomposed per row per input parameter a or b.

NUM_SELECTORS

Number of selector columns in the trace.

OP_CYCLE_LEN

The number of rows required to compute an operation in the Bitwise chiplet.

OUTPUT_COL_IDX

The index of the column containing the aggregated output value within the bitwise chiplet execution trace.

PREV_OUTPUT_COL_IDX

The index of the column containing the aggregated output value within the bitwise chiplet execution trace.

TRACE_WIDTH

Number of columns needed to record an execution trace of the bitwise chiplet.

Type Aliases

Selectors