miden_air::trace::chiplets

Module kernel_rom

Constants

KERNEL_PROC_CALL_LABEL

Specifies a kernel procedure call operation to access a procedure in the kernel ROM.

KERNEL_PROC_INIT_LABEL

Specified the label of the kernel ROM initialization request by the verifier.

TRACE_WIDTH

Number of columns needed to record an execution trace of the kernel ROM chiplet.