miden_air::trace

Module decoder

Constants

ADDR_COL_IDX

Index of the column holding code block IDs (which are row addresses from the hasher table).

DECODER_OP_BITS_OFFSET

DECODER_USER_OP_HELPERS_OFFSET

GROUP_COUNT_COL_IDX

Index of the operation group count column in the decoder trace.

HASHER_STATE_OFFSET

Index at which hasher state columns start in the decoder trace.

HASHER_STATE_RANGE

Location of hasher columns in the decoder trace.

IN_SPAN_COL_IDX

Index of the in_span column in the decoder trace.

IS_CALL_FLAG_COL_IDX

Index of a flag column which indicates whether an ending block is a CALL or DYNCALL block.

IS_LOOP_BODY_FLAG_COL_IDX

Index of a flag column which indicates whether an ending block is a body of a loop.

IS_LOOP_FLAG_COL_IDX

Index of a flag column which indicates whether an ending block is a LOOP block.

IS_SYSCALL_FLAG_COL_IDX

Index of a flag column which indicates whether an ending block is a SYSCALL block.

NUM_HASHER_COLUMNS

Number of hasher columns in the decoder trace.

NUM_OP_BATCH_FLAGS

Number of operation batch flag columns.

NUM_OP_BITS

Number of columns needed to hold a binary representation of opcodes.

NUM_OP_BITS_EXTRA_COLS

Number of columns needed for degree reduction of the operation flags.

NUM_USER_OP_HELPERS

Number of helper registers available to user ops.

OP_BATCH_1_GROUPS

Operation batch consists of 1 operation group.

OP_BATCH_2_GROUPS

Operation batch consists of 2 operation groups.

OP_BATCH_4_GROUPS

Operation batch consists of 4 operation groups.

OP_BATCH_8_GROUPS

Operation batch consists of 8 operation groups.

OP_BATCH_FLAGS_OFFSET

Index at which operation batch flag columns start in the decoder trace.

OP_BATCH_FLAGS_RANGE

Location of operation batch flag columns in the decoder trace.

OP_BITS_EXTRA_COLS_OFFSET

Index at which the op bits extra columns start in the decoder trace.

OP_BITS_EXTRA_COLS_RANGE

Location of the operation bits extra columns (for degree reduction) in the decoder trace.

OP_BITS_OFFSET

Index at which operation bit columns start in the decoder trace.

OP_BITS_RANGE

Location of operation bits columns in the decoder trace.

OP_INDEX_COL_IDX

Index of the operation index column in the decoder trace.

P1_COL_IDX

Running product column representing block stack table.

P2_COL_IDX

Running product column representing block hash table

P3_COL_IDX

Running product column representing op group table.

USER_OP_HELPERS_OFFSET

Index at which helper registers available to user ops start. The first two helper registers are used by the decoder itself.