miden_air::trace

Module range

Constants

B_RANGE_COL_IDX

The running product column used for verifying that the range check lookups performed in the Stack and the Memory chiplet match the values checked in the Range Checker.

M_COL_IDX

A column to hold the multiplicity of how many times the value is being range-checked.

V_COL_IDX

A column to hold the values being range-checked.