miden_air::trace

Module stack

Constants

B0_COL_IDX

Index of the b0 helper column in the stack trace. This column holds the current stack depth.

B1_COL_IDX

Index of the b1 helper column in the stack trace. This column holds the address of the top item in the stack overflow table.

H0_COL_IDX

Index of the h0 helper column in the stack trace. This column contains 1 / (b0 - 16) when b0 != 16, and ZERO otherwise.

NUM_STACK_HELPER_COLS

Number of bookkeeping and helper columns in the stack trace.

STACK_TOP_OFFSET

Index at which stack item columns start in the stack trace.

STACK_TOP_RANGE

Location of stack top items in the stack trace.