miden_air::trace

Struct MainTrace

pub struct MainTrace { /* private fields */ }

Implementations

impl MainTrace

pub fn new(main_trace: ColMatrix<Felt>, last_program_row: RowIndex) -> Self

pub fn to_row_major(&self) -> RowMajorMatrix<Felt>

Converts this column-major MainTrace into a row-major matrix.

pub fn num_rows(&self) -> usize

pub fn last_program_row(&self) -> RowIndex

pub fn get_column_range(&self, range: Range<usize>) -> Vec<Vec<Felt>>

pub fn clk(&self, i: RowIndex) -> Felt

Returns the value of the clk column at row i.

pub fn ctx(&self, i: RowIndex) -> Felt

Returns the value of the ctx column at row i.

pub fn addr(&self, i: RowIndex) -> Felt

Returns the value in the block address column at the row i.

pub fn is_addr_change(&self, i: RowIndex) -> bool

Helper method to detect change of address.

pub fn helper_register(&self, i: usize, row: RowIndex) -> Felt

The i-th decoder helper register at row.

pub fn decoder_hasher_state(&self, i: RowIndex) -> [Felt; 8]

Returns the hasher state at row i.

pub fn decoder_hasher_state_first_half(&self, i: RowIndex) -> Word

Returns the first half of the hasher state at row i.

pub fn decoder_hasher_state_second_half(&self, i: RowIndex) -> Word

Returns the second half of the hasher state at row i.

pub fn decoder_hasher_state_element(&self, element: usize, i: RowIndex) -> Felt

Returns a specific element from the hasher state at row i.

pub fn fn_hash(&self, i: RowIndex) -> [Felt; 4]

Returns the current function hash (i.e., root) at row i.

pub fn is_loop_body_flag(&self, i: RowIndex) -> Felt

Returns the is_loop_body flag at row i.

pub fn is_loop_flag(&self, i: RowIndex) -> Felt

Returns the is_loop flag at row i.

pub fn is_call_flag(&self, i: RowIndex) -> Felt

Returns the is_call flag at row i.

pub fn is_syscall_flag(&self, i: RowIndex) -> Felt

Returns the is_syscall flag at row i.

pub fn op_batch_flag(&self, i: RowIndex) -> [Felt; 3]

Returns the operation batch flags at row i. This indicates the number of op groups in the current batch that is being processed.

pub fn group_count(&self, i: RowIndex) -> Felt

Returns the operation group count. This indicates the number of operation that remain to be executed in the current span block.

pub fn delta_group_count(&self, i: RowIndex) -> Felt

Returns the delta between the current and next group counts.

pub fn is_in_span(&self, i: RowIndex) -> Felt

Returns the in_span flag at row i.

pub fn get_op_code(&self, i: RowIndex) -> Felt

Constructs the i-th op code value from its individual bits.

pub fn row_iter(&self) -> impl Iterator<Item = RowIndex>

Returns an iterator of RowIndex values over the row indices of this trace.

pub fn is_left_shift(&self, i: RowIndex) -> bool

Returns a flag indicating whether the current operation induces a left shift of the operand stack.

pub fn is_right_shift(&self, i: RowIndex) -> bool

Returns a flag indicating whether the current operation induces a right shift of the operand stack.

pub fn stack_depth(&self, i: RowIndex) -> Felt

Returns the value of the stack depth column at row i.

pub fn stack_element(&self, column: usize, i: RowIndex) -> Felt

Returns the element at row i in a given stack trace column.

pub fn stack_word(&self, start: usize, i: RowIndex) -> Word

Returns a word from the stack starting at start index at row i, in LE order.

The word is read such that word[0] comes from stack position start (top), word[1] from start + 1, etc.

pub fn parent_overflow_address(&self, i: RowIndex) -> Felt

Returns the address of the top element in the stack overflow table at row i.

pub fn is_non_empty_overflow(&self, i: RowIndex) -> bool

Returns a flag indicating whether the overflow stack is non-empty.

pub fn chiplet_selector_0(&self, i: RowIndex) -> Felt

Returns chiplet column number 0 at row i.

pub fn chiplet_selector_1(&self, i: RowIndex) -> Felt

Returns chiplet column number 1 at row i.

pub fn chiplet_selector_2(&self, i: RowIndex) -> Felt

Returns chiplet column number 2 at row i.

pub fn chiplet_selector_3(&self, i: RowIndex) -> Felt

Returns chiplet column number 3 at row i.

pub fn chiplet_selector_4(&self, i: RowIndex) -> Felt

Returns chiplet column number 4 at row i.

pub fn chiplet_selector_5(&self, i: RowIndex) -> Felt

Returns chiplet column number 5 at row i.

pub fn is_hash_row(&self, i: RowIndex) -> bool

Returns true if a row is part of the hash chiplet.

pub fn chiplet_hasher_state(&self, i: RowIndex) -> [Felt; 12]

Returns the (full) state of the hasher chiplet at row i.

pub fn chiplet_node_index(&self, i: RowIndex) -> Felt

Returns the hasher’s node index column at row i

pub fn is_bitwise_row(&self, i: RowIndex) -> bool

Returns true if a row is part of the bitwise chiplet.

pub fn chiplet_bitwise_a(&self, i: RowIndex) -> Felt

Returns the bitwise column holding the aggregated value of input a at row i.

pub fn chiplet_bitwise_b(&self, i: RowIndex) -> Felt

Returns the bitwise column holding the aggregated value of input b at row i.

pub fn chiplet_bitwise_z(&self, i: RowIndex) -> Felt

Returns the bitwise column holding the aggregated value of the output at row i.

pub fn is_memory_row(&self, i: RowIndex) -> bool

Returns true if a row is part of the memory chiplet.

pub fn chiplet_memory_ctx(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing memory context.

pub fn chiplet_memory_word(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing memory address.

pub fn chiplet_memory_idx0(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing 0th bit of the word index.

pub fn chiplet_memory_idx1(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing 1st bit of the word index.

pub fn chiplet_memory_clk(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing clock cycle.

pub fn chiplet_memory_value_0(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing the zeroth memory value element.

pub fn chiplet_memory_value_1(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing the first memory value element.

pub fn chiplet_memory_value_2(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing the second memory value element.

pub fn chiplet_memory_value_3(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing the third memory value element.

pub fn is_ace_row(&self, i: RowIndex) -> bool

Returns true if a row is part of the ACE chiplet.

pub fn chiplet_ace_start_selector(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_block_selector(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_ctx(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_ptr(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_clk(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_eval_op(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_num_eval_rows(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_id_0(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_v_0_0(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_v_0_1(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_wire_0(&self, i: RowIndex) -> [Felt; 3]

pub fn chiplet_ace_id_1(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_v_1_0(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_v_1_1(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_wire_1(&self, i: RowIndex) -> [Felt; 3]

pub fn chiplet_ace_id_2(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_v_2_0(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_v_2_1(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_wire_2(&self, i: RowIndex) -> [Felt; 3]

pub fn chiplet_ace_m_1(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_m_0(&self, i: RowIndex) -> Felt

pub fn chiplet_ace_is_read_row(&self, i: RowIndex) -> bool

pub fn chiplet_ace_is_eval_row(&self, i: RowIndex) -> bool

pub fn is_kernel_row(&self, i: RowIndex) -> bool

Returns true if a row is part of the kernel chiplet.

pub fn chiplet_kernel_is_first_hash_row(&self, i: RowIndex) -> bool

Returns true when the i-th row of the s_first column in the kernel chiplet is one, i.e., when this is the first row in a range of rows containing the same kernel proc hash.

pub fn chiplet_kernel_root_0(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing the zeroth element of the kernel procedure root.

pub fn chiplet_kernel_root_1(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing the first element of the kernel procedure root.

pub fn chiplet_kernel_root_2(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing the second element of the kernel procedure root.

pub fn chiplet_kernel_root_3(&self, i: RowIndex) -> Felt

Returns the i-th row of the chiplet column containing the third element of the kernel procedure root.

pub fn f_mv(&self, i: RowIndex) -> bool

Returns true if the hasher chiplet flags indicate the initialization of verifying a Merkle path to an old node during Merkle root update procedure (MRUPDATE).

pub fn f_mva(&self, i: RowIndex) -> bool

Returns true if the hasher chiplet flags indicate the continuation of verifying a Merkle path to an old node during Merkle root update procedure (MRUPDATE).

pub fn f_mu(&self, i: RowIndex) -> bool

Returns true if the hasher chiplet flags indicate the initialization of verifying a Merkle path to a new node during Merkle root update procedure (MRUPDATE).

pub fn f_mua(&self, i: RowIndex) -> bool

Returns true if the hasher chiplet flags indicate the continuation of verifying a Merkle path to a new node during Merkle root update procedure (MRUPDATE).