miden_air

Trait PermutationAirBuilder

pub trait PermutationAirBuilder: ExtensionBuilder {
    type MP: WindowAccess<Self::VarEF>;
    type RandomVar: Into<Self::ExprEF> + Copy;
    type PermutationVar: Into<Self::ExprEF> + Clone;

    // Required methods
    fn permutation(&self) -> Self::MP;
    fn permutation_randomness(&self) -> &[Self::RandomVar];
    fn permutation_values(&self) -> &[Self::PermutationVar];
}

Trait for builders supporting permutation arguments (e.g., for lookup constraints).

Required Associated Types

type MP: WindowAccess<Self::VarEF>

Two-row window over the permutation trace columns.

type RandomVar: Into<Self::ExprEF> + Copy

Randomness variable type used in permutation commitments.

type PermutationVar: Into<Self::ExprEF> + Clone

Value type for expected cumulated values used in global lookup arguments.

Required Methods

fn permutation(&self) -> Self::MP

Return the current and next row slices of the permutation trace.

fn permutation_randomness(&self) -> &[Self::RandomVar]

Return the list of randomness values for permutation argument.

fn permutation_values(&self) -> &[Self::PermutationVar]

Return the expected cumulated values for global lookup arguments.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.